requirements.txt
# Core
python-dotenv>=1.0.0
requests>=2.31.0

# LiveKit Agent Stack
livekit-agents>=1.3.12
livekit-plugins-google>=1.0.0
livekit-plugins-noise-cancellation>=1.0.0

# Memory (mem0)
mem0ai>=0.1.15

# Search
langchain-community>=0.2.0
duckduckgo-search>=5.3.0

# Timezone support (CRITICAL for time tool)
tzdata>=2024.1

# Email
secure-smtplib>=0.1.1
